Map `Identifier.URI` from apparatus dublin core to nothing.
As `Identifier.URI` points to a respective article's abstract page (where its DOI is),
it is kinda uncertain whether the *official website* property (`P856`) is an appropriate counterpart.
Also, `P856` is not listed in the *WikiProject Source Metadata* [proposal for bibliographic
properties](https://www.wikidata.org/wiki/Template:Bibliographic_properties) and we should stick to that
list for the sake of clean methodology.